:recycle: Home page refactoring
